--- Brasa/noticias/home.html
@@ -1,122 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Adiciona o CSRF Token para segurança nas requisições -->
-    <meta name="csrf-token" content="{{ csrf_token }}">
-    <title>Jornal BRASA - Últimas Notícias</title>
-    <style>
-        body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif; line-height: 1.6; margin: 0; padding: 20px; background-color: #f4f4f4; }
-        .container { max-width: 800px; margin: auto; background: #fff; padding: 20px; border-radius: 8px; }
-        .header { display: flex; justify-content: space-between; align-items: center; border-bottom: 2px solid #eee; padding-bottom: 10px; }
-        h1 { color: #333; margin: 0; }
-        .header a { text-decoration: none; background-color: #007bff; color: white; padding: 8px 15px; border-radius: 5px; font-size: 0.9em; }
-        .noticia { position: relative; border-bottom: 1px solid #ddd; padding: 20px 0; margin-bottom: 20px; }
-        .noticia h2 { color: #c0392b; margin-top: 0; }
-        .noticia-meta { font-size: 0.9em; color: #666; margin-bottom: 10px; }
-        .noticia-conteudo { text-align: justify; }
-
-        /* Estilo do botão Salvar */
-        .save-btn {
-            position: absolute;
-            top: 20px;
-            right: 0;
-            background: none;
-            border: 1px solid #ccc;
-            border-radius: 50%;
-            width: 40px;
-            height: 40px;
-            cursor: pointer;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            transition: background-color 0.2s, border-color 0.2s;
-        }
-        .save-btn svg {
-            width: 20px;
-            height: 20px;
-            fill: #333;
-            transition: fill 0.2s;
-        }
-        .save-btn:hover {
-            border-color: #888;
-        }
-        /* Estilo do botão QUANDO ESTÁ SALVO */
-        .save-btn.salvo {
-            background-color: #ffdd57; /* Amarelo */
-            border-color: #ffdd57;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="header">
-            <h1>Jornal BRASA</h1>
-            {% if user.is_authenticated %}
-                <a href="{% url 'noticias_salvas' %}">Meus Itens Salvos</a>
-            {% else %}
-                <a href="{% url 'admin:login' %}">Login</a>
-            {% endif %}
-        </div>
-        
-        {% for noticia in noticias %}
-            <article class="noticia">
-                <!-- Botão para Salvar -->
-                {% if user.is_authenticated %}
-                    <button 
-                        class="save-btn {% if noticia.id in noticias_salvas_ids %}salvo{% endif %}" 
-                        data-id="{{ noticia.id }}" 
-                        title="Salvar notícia">
-                        <svg viewBox="0 0 24 24"><path d="M17 3H7c-1.1 0-1.99.9-1.99 2L5 21l7-3 7 3V5c0-1.1-.9-2-2-2z"></path></svg>
-                    </button>
-                {% endif %}
-
-                <h2>{{ noticia.titulo }}</h2>
-                <div class="noticia-meta">
-                    Por <strong>{{ noticia.autor }}</strong> em {{ noticia.data_publicacao|date:"d/m/Y H:i" }} | Categoria: {{ noticia.get_categoria_display }}
-                </div>
-                <div class="noticia-conteudo">
-                    {{ noticia.conteudo|linebreaks }}
-                </div>
-            </article>
-        {% empty %}
-            <p>Nenhuma notícia foi publicada ainda.</p>
-        {% endfor %}
-    </div>
-
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            // Pega o CSRF token do meta tag para segurança
-            const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
-
-            document.querySelectorAll('.save-btn').forEach(button => {
-                button.addEventListener('click', function() {
-                    const noticiaId = this.dataset.id;
-                    const url = `/salvar/${noticiaId}/`; // A URL que configuramos no Passo 4
-
-                    // Envia a requisição para o servidor
-                    fetch(url, {
-                        method: 'POST',
-                        headers: {
-                            'X-CSRFToken': csrfToken,
-                            'Content-Type': 'application/json'
-                        }
-                    })
-                    .then(response => response.json())
-                    .then(data => {
-                        // A resposta do servidor dirá se a notícia está salva (true) ou não (false)
-                        if (data.salvo) {
-                            this.classList.add('salvo'); // Adiciona a classe para ficar amarelo
-                        } else {
-                            this.classList.remove('salvo'); // Remove a classe
-                        }
-                    })
-                    .catch(error => console.error('Erro:', error));
-                });
-            });
-        });
-    </script>
-</body>
-</html>
-
